+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unrealeased
 - [Remove] 移除一切附属 package 并停止开发附属 package
+- [Fix] 修复了在将 setter `ISessionContext#timeout` 设为大于 `Number.MAX_SAFE_INT` 值时仍然运行 `setTimeout()` 的问题
+- [Add] 增加了导出变量 `sessionCount`，为正在进行的会话计数。
 
 ## 0.4.0
 - [Change] 将处理队列的超时在 `init({...})` 中更名为 `middlewareTimeout`